fix(cli): remove the git-root lru_cache in _paths.py entirely
Three review rounds each found another env var (GIT_DIR, GIT_WORK_TREE,
GIT_COMMON_DIR, then GIT_CEILING_DIRECTORIES, GIT_DISCOVERY_ACROSS_FILESYSTEM,
GIT_IMPLICIT_WORK_TREE) that changes `git rev-parse --show-toplevel`'s
answer for a fixed cwd. The invalidation domain can't be enumerated safely,
and the callers (plugins/discovery.py, cli/_providers.py,
cli/orchestrate/__init__.py, state/provenance.py) only call
find_lionagi_dirs() a handful of times per process at single-digit-ms cost.

find_lionagi_dirs() now calls _find_git_root(cwd) directly, no memoization.
Removes _find_git_root_cached, _git_context, _GIT_CONTEXT_VARS, and the
find_lionagi_dirs.cache_clear hook. clear_lionagi_dirs_cache is private-only
(never re-exported from lionagi/__init__.py or any public package) so it is
removed cleanly with a CHANGELOG Removed entry rather than a deprecation
cycle.

Retargets tests/cli/test_lionagi_dir_cache_invalidation.py to
test_lionagi_dir_discovery.py, dropping the cache-keying and repeated-call
cache-hit assertions that no longer apply, keeping the topology-visibility
coverage and the GIT_DIR/GIT_WORK_TREE redirect regression test. Updates the
two other test files that called the now-removed clear_lionagi_dirs_cache.
